Split creator tests into multiple files
